renamed bcrypt to auth and test create_access_token
